+++ src/frontend/Komponenter/Behandling/Vurdering/HjemmelSelect.module.css
@@ -1,0 +1,4 @@
+.container {
+    margin: 2rem 4rem 2rem 4rem;
+    width: 18rem;
+}
